refactor: 인증 서비스와 조회 서비스 분리
