+++ backend/src/main/java/harustudy/backend/admin/controller/AdminController.java
@@ -1,8 +1,8 @@
 package harustudy.backend.admin.controller;
 
 import harustudy.backend.admin.dto.*;
+import harustudy.backend.admin.service.AdminAuthService;
 import harustudy.backend.admin.service.AdminService;
-import harustudy.backend.study.dto.StudiesResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -19,12 +19,13 @@ import java.util.UUID;
 @RestController
 public class AdminController {
 
+    private final AdminAuthService adminAuthService;
     private final AdminService adminService;
 
     @PostMapping(value = "/login", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
     public ResponseEntity<Void> login(AdminLoginRequest request,
                                       HttpServletRequest httpServletRequest) {
-        UUID sessionId = adminService.login(request);
+        UUID sessionId = adminAuthService.login(request);
         HttpSession session = httpServletRequest.getSession();
         session.setAttribute("SESSION", sessionId);
         session.setMaxInactiveInterval(3600);
